refactor: Perform codebase refactor to separate appropriate functions
