+++ phylo2vec/src/tree_vec/ops/mod.rs
@@ -1,8 +1,18 @@
 pub mod avl;
+pub mod newick;
 pub mod vector;
 
-#[allow(unused_imports)]
+use crate::tree_vec::types::Ancestry;
+
 pub use vector::{
-    build_newick, build_vector, find_coords_of_first_leaf, get_ancestry, get_pairs, get_pairs_avl,
-    order_cherries, order_cherries_no_parents, to_newick, Ancestry,
+    build_vector, find_coords_of_first_leaf, get_ancestry, get_pairs, get_pairs_avl,
+    order_cherries, order_cherries_no_parents,
 };
+
+pub use newick::build_newick;
+
+/// Recover a rooted tree (in Newick format) from a Phylo2Vec vector
+pub fn to_newick(v: &Vec<usize>) -> String {
+    let ancestry: Ancestry = get_ancestry(&v);
+    build_newick(&ancestry)
+}
